deployment-services/requirements.txt
asgiref==3.11.0
awscrt==0.31.2
boto3==1.42.59
botocore==1.42.59
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
Django==5.2.10
django-filter==25.2
djangorestframework==3.16.1
dotenv==0.9.9
idna==3.11
jmespath==1.1.0
Markdown==3.10.1
pika==1.3.2
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
rabbitmq==0.2.0
redis==5.0.1
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
uuid6==2024.7.10
python-json-logger==3.2.1
prometheus-client==0.21.1
drf-spectacular==0.28.0
